slightly refactor forecast downloader
